A checkpoint commit of two BAM reading projects going on simultaneously.  These two projects are works in progress, and this checkin will provide a baseline against which to gauge improvements to both projects.
Low-memory BAM protoshards (disabled by default):
- Currently passing ValidatingPileupIntegrationTest.
- Gets progressively slower throughout the traversal, but should run at least as fast as original implementation.
- Uses 10+ file handles per BAM, but should use 3.

BAM performance microbenchmark test system:
- Currently tests performance of BAM reading using SAM-JDK vs. GATK
- Tests do not hit all GATK performance hotspots.
- New tests that require input data in a slightly different form are hard to implement.
- Output of test results is not easily parseable (investigating Google Caliper for possible improvements).



git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@5317 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/gatk/datasources/reads/IntervalSharder.java
@@ -158,8 +158,8 @@ public class IntervalSharder {
             if(!binIterator.hasNext())
                 break;
 
-            int locationStart = (int)location.getStart();
-            final int locationStop = (int)location.getStop();
+            int locationStart = location.getStart();
+            final int locationStop = location.getStop();
 
             // Advance to first bin.
             while(binIterator.peek().stop < locationStart)
